chore: migrate client and docs to Rstack CLI
