+++ packages/client/config/constants.ts
@@ -1,18 +1,21 @@
 import { Constants } from '@rsdoctor/shared/types';
 
-import path from 'path';
+import path from 'node:path';
 
-export const DistPath = path.resolve(__dirname, '../dist');
+export const DistPath = path.resolve(import.meta.dirname, '../dist');
 
-export const DistResourcePath = path.resolve(__dirname, '../dist/resource');
+export const DistResourcePath = path.resolve(
+  import.meta.dirname,
+  '../dist/resource',
+);
 
 export const WebpackRsdoctorDirPath = path.resolve(
-  __dirname,
+  import.meta.dirname,
   `../dist/${Constants.RsdoctorOutputFolder}`,
 );
 
 export const WebpackStatsFilePath = path.resolve(
-  __dirname,
+  import.meta.dirname,
   '../dist/stats.json',
 );
 
@@ -20,4 +23,7 @@ export const PortForWeb = 8681;
 
 export const PortForCLI = 8123;
 
-export const ClientEntry = path.resolve(__dirname, '../src/index.tsx');
+export const ClientEntry = path.resolve(
+  import.meta.dirname,
+  '../src/index.tsx',
+);
